requirements.txt
backcall==0.2.0
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.1
django-crispy-forms==1.9.2
django-crum==0.7.7
django-debug-toolbar==2.2
django-environ==0.4.5
django-extensions==3.0.5
django-import-export==2.3.0
django-model-utils==4.0.0
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
ipython==7.16.1
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.17.2
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.4
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.6
psycopg2==2.8.5
ptyprocess==0.6.0
Pygments==2.6.1
PyJWT==1.7.1
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
traitlets==4.3.3
wcwidth==0.2.5
xlrd==1.2.0
xlwt==1.3.0
